Use broadcast where needed. Implement urgent mutex lock

+++ src/cli/command/RemoveFileCommand.java
@@ -17,7 +17,9 @@ public class RemoveFileCommand implements CLICommand {
     @Override
     public void execute(String fileName) {
         // lock
-        AppConfig.chordState.mutex.lock(AppConfig.chordState.getAllNodeInfo().stream().map(ServentInfo::getListenerPort).collect(Collectors.toSet()));
+        AppConfig.chordState.mutex.lock();
+        if(!AppConfig.isAlive.get())
+            return;
 
         // delete
         AppConfig.chordState.deleteValue(fileName, AppConfig.myServentInfo.getListenerPort());
